pyproject.toml
[project]
name = "qdrant-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "qdrant-client>=1.13.2",
    "requests>=2.32.3",
    "streamlit>=1.42.0",
    "torch>=2.6.0",
    "tqdm>=4.67.1",
    "transformers>=4.48.3",
    "watchdog>=6.0.0",
]
